feat(website): glass panel + real signers on the certificate
One centered translucent panel carries all certificate content over the
full-bleed space photo (alpha-ramped tint thinnest over the sunrise,
deepest under the code/QR row); full-bleed scrims reduced to a vignette
so the photo regains vibrancy. Signers: Felipe Salinas (CEO), Julian
Arango (CTO). Matching panel language on the OG card.

Fixes HOU-1173
