+++ website/lib/certs/template-og.mjs
@@ -1,41 +1,50 @@
+import { backdrop } from "./backdrop.mjs";
 import {
   BODY_FONT,
-  backdrop,
   HAIRLINE,
   INK,
   INK_MUTED,
   INK_SUBTLE,
   lockup,
-  NAVY,
   navy,
   text,
 } from "./chrome.mjs";
 import { certCopy } from "./copy.mjs";
 import h from "./h.mjs";
+import { canvas, panel, tint } from "./panel.mjs";
 
 /**
- * The social card: 1200x630 on the same photograph as the certificate.
+ * The social card: 1200x630 on the same photograph and the same glass panel as
+ * the certificate.
  *
  * Deliberately NOT a shrunken certificate — at feed size a document reads as
- * grey noise. The card keeps the mark, the claim, the name and the event, all
- * left-aligned in the dark half of the sky, and hands the bottom third back to
- * the photograph: at this size the sunrise IS the design.
+ * grey noise. The card keeps the panel language for family recognition, but
+ * everything inside it stays left-aligned and set at feed size: the mark, the
+ * claim, the name and the event, with the sunrise coming up through the lower
+ * half of the glass behind them. The inset is tighter than the certificate's,
+ * proportionally, because at thumbnail size a wide frame eats the type.
  */
 export const OG_WIDTH = 1200;
 export const OG_HEIGHT = 630;
 
+/** Photograph left visible on all four sides of the panel. */
+const INSET = 40;
+
 /**
- * Two scrims, because the card's type is left-aligned rather than centred.
- *
- * The vertical one holds the sky down for the header and the claim and releases
- * before the horizon. The horizontal one deepens only the left, which is what
- * lets the name sit on a photograph without a panel behind it — and it leaves
- * the sunrise and the far ship untouched on the right.
+ * Same shape of ramp as the certificate's, one notch lighter throughout: the
+ * card is read at thumbnail size, where a heavy tint is the difference between
+ * a photograph and a grey rectangle in a feed.
  */
-const SCRIM = [
-  `linear-gradient(180deg, ${navy(0.7)} 0%, ${navy(0.58)} 38%, ${navy(0.26)} 62%, ${navy(0)} 78%)`,
-  `linear-gradient(90deg, ${navy(0.52)} 0%, ${navy(0.28)} 44%, ${navy(0)} 70%)`,
-];
+const PANEL = tint([
+  [0.42, 0],
+  [0.36, 36],
+  [0.32, 62],
+  [0.4, 82],
+  [0.52, 100],
+]);
+
+/** A vignette, nothing more — the panel carries the contrast. */
+const VIGNETTE = `linear-gradient(180deg, ${navy(0.1)} 0%, ${navy(0.01)} 28%, ${navy(0)} 62%, ${navy(0.12)} 100%)`;
 
 /** One notch smaller than the printed certificate's ladder. */
 function nameSize(name) {
@@ -55,29 +64,16 @@ export function ogCardElement(item) {
 
   return h(
     "div",
-    {
-      style: {
-        display: "flex",
-        flexDirection: "column",
-        position: "relative",
-        overflow: "hidden",
+    { style: canvas(OG_WIDTH, OG_HEIGHT, BODY_FONT) },
+    ...backdrop(OG_WIDTH, OG_HEIGHT, { focusY: 0.5, scrim: VIGNETTE }),
+    panel(
+      {
         width: OG_WIDTH,
         height: OG_HEIGHT,
-        backgroundColor: NAVY,
-        fontFamily: BODY_FONT,
-      },
-    },
-    ...backdrop(OG_WIDTH, OG_HEIGHT, { focusY: 0.5, scrim: SCRIM }),
-    h(
-      "div",
-      {
-        style: {
-          display: "flex",
-          flexDirection: "column",
-          width: "100%",
-          height: "100%",
-          padding: "54px 68px",
-        },
+        inset: INSET,
+        radius: 22,
+        padding: "38px 48px 46px 48px",
+        background: PANEL,
       },
       // ── Issuer left, credential metadata right ───────────────────────────
       h(
@@ -91,17 +87,17 @@ export function ogCardElement(item) {
             width: "100%",
           },
         },
-        lockup({ helmet: 46, word: 22, domain: 13 }),
+        lockup({ helmet: 44, word: 21, domain: 12 }),
         text(meta, {
-          fontSize: 19,
+          fontSize: 18,
           fontWeight: 400,
           letterSpacing: 1.4,
           color: INK_SUBTLE,
         }),
       ),
       // ── The claim ────────────────────────────────────────────────────────
-      // Bottom-anchored inside the sky: a name long enough to wrap grows UP
-      // into the dark, never down onto the sunrise.
+      // Bottom-anchored on the glass: a name long enough to wrap grows UP into
+      // the dark half of the panel, never down onto the sunrise.
       h(
         "div",
         {
@@ -111,8 +107,7 @@ export function ogCardElement(item) {
             flex: 1,
             justifyContent: "flex-end",
             alignItems: "flex-start",
-            width: 880,
-            paddingBottom: 118,
+            width: "100%",
           },
         },
         text(copy.certificateOf, {
@@ -137,7 +132,7 @@ export function ogCardElement(item) {
           color: INK,
         }),
         text(item.eventTitle, {
-          marginTop: 16,
+          marginTop: 14,
           width: "100%",
           fontSize: 30,
           fontWeight: 600,
